notification-app/notificationAPI/requirements.txt
asgiref==3.5.2
Django==4.1.4
django-braces==1.15.0
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
